fix: move chunk-helpers inside src/ to fix Docker build
The chunk-helpers module was in test/ which is outside tsconfig's
rootDir (src/), causing tsc to fail during Docker image builds.
